move shuffleArray into a module function
