+++ src/components/Question.tsx
@@ -1,21 +1,13 @@
 import { useState } from 'react';
 import parse from 'html-react-parser';
 import Answer from './Answer';
+import shuffleArray from '../functions/shuffleArray';
 
 interface QuestionProps {
   question: string;
   answers: string[];
   corrrectAnswer: string;
 }
-
-const shuffleArray = (array: string[]) => {
-  const shuffledArray = [...array];
-  for (let i = shuffledArray.length - 1; i > 0; i--) {
-    const j = Math.floor(Math.random() * (i + 1));
-    [shuffledArray[i], shuffledArray[j]] = [shuffledArray[j], shuffledArray[i]];
-  }
-  return shuffledArray;
-};
 
 const Question = ({ question, answers }: QuestionProps) => {
   const [selectedAnswer, setSelectedAnswer] = useState('');
